% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{redmine_add_watcher}
\alias{redmine_add_watcher}
\title{Add watcher to an issue}
\usage{
redmine_add_watcher(issue_id, user_id)
}
\arguments{
\item{issue_id}{Issue id}

\item{user_id}{User id to add as a watcher}
}
\description{
Add watcher to an issue
}
\author{
Maxim Nazarov
}
