% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{redmine_create_category}
\alias{redmine_create_category}
\title{Create category}
\usage{
redmine_create_category(project_id, name, assigned_to_id = NULL, ...)
}
\arguments{
\item{project_id}{Project id to add issue to}

\item{name}{Category name}

\item{assigned_to_id}{User id to assign to}

\item{...}{Further parameters}
}
\value{
id of the created category
}
\description{
Create category
}
\examples{
\dontrun{
# first get necessary ids
projectID <- redmine_search_id("testProject", 
    endpoint = "projects")

# now create category
issueID <- redmine_create_category(
   project_id = projectID, 
   name = "Functionality")
}
}
\author{
Laure Cougnaud
}
