% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{redmine_create_issue}
\alias{redmine_create_issue}
\alias{redmine_update_issue}
\alias{redmine_delete_issue}
\title{Create/update/delete issue}
\usage{
redmine_create_issue(project_id, subject, description = NULL, files = NULL,
  tracker_id = NULL, status_id = NULL, priority_id = NULL,
  category_id = NULL, fixed_version_id = NULL, assigned_to_id = NULL,
  parent_issue_id = NULL, custom_fields = NULL, watcher_user_ids = NULL,
  is_private = NULL, estimated_hours = NULL, ...)

redmine_update_issue(issue_id, notes = NULL, project_id = NULL,
  tracker_id = NULL, status_id = NULL, subject = NULL,
  private_notes = FALSE, files = NULL, ...)

redmine_delete_issue(issue_id)
}
\arguments{
\item{project_id}{Project id to add issue to}

\item{subject}{Issue subject (title)}

\item{description}{Issue description}

\item{files}{List of filenames to add (upload) to the issue}

\item{tracker_id}{Tracker id(s)}

\item{status_id}{Status id}

\item{priority_id}{Priority id}

\item{category_id}{Issue category id}

\item{fixed_version_id}{Target Version id}

\item{assigned_to_id}{User id to assign to}

\item{parent_issue_id}{Parent issue id}

\item{custom_fields}{Custom fields, as list, see examples}

\item{watcher_user_ids}{User id(s) to add as watchers}

\item{is_private}{Whether the issue is private}

\item{estimated_hours}{Estimated number of hours}

\item{...}{Further parameters}

\item{issue_id}{Issue id}

\item{notes}{Notes (comments) to add}

\item{private_notes}{Private notes to add}
}
\value{
id of the created issue for \code{redmine_create_issue}
}
\description{
These functions implement Redmine API calls to work with issues.
}
\examples{
\dontrun{
# first get necessary ids
projectID <- redmine_search_id("testProject", 
    endpoint = "projects")

urgentID <- redmine_search_id("Urgent", 
    endpoint = "issue_priorities")

# now create issue
issueID <- redmine_create_issue(
    project_id = projectID, 
    subject = "Urgent task", 
    description = "Do it quick!", 
    priority_id = urgentID,
    files = "details.txt")

# modify using custom fields
redmine_update_issue(issueID, 
    custom_fields = list(
        list(id = 1, name = "myField", value = 100),
        list(id = 2, name = "myField2", value = 200)
    )
)   
}
}
\seealso{
\url{http://www.redmine.org/projects/redmine/wiki/Rest_Issues}
}
\author{
Maxim Nazarov
}
