% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{redmine_create_project}
\alias{redmine_create_project}
\alias{redmine_update_project}
\alias{redmine_delete_project}
\title{Create/update/delete project}
\usage{
redmine_create_project(name, identifier, description = NULL,
  homepage = NULL, is_public = TRUE, parent_id = NULL,
  inherit_members = TRUE, tracker_ids = NULL,
  enabled_module_names = c("boards", "calendar", "documents", "files",
  "gantt", "issue_tracking", "news", "repository", "time_tracking",
  "wiki"), ...)

redmine_update_project(project_id, name = NULL, identifier = NULL,
  description = NULL, homepage = NULL, is_public = NULL,
  parent_id = NULL, inherit_members = NULL, tracker_ids = NULL,
  enabled_module_names = NULL, ...)

redmine_delete_project(project_id)
}
\arguments{
\item{name}{Project name}

\item{identifier}{Project identifier (no spaces)}

\item{description}{Project description}

\item{homepage}{Project homepage}

\item{is_public}{Whether project should be made public (TRUE by default)}

\item{parent_id}{Project's parent id}

\item{inherit_members}{Whether to inherit members from parent}

\item{tracker_ids}{Trackers id's}

\item{enabled_module_names}{Enabled modules, a subset of \code{c("boards", 
"calendar", "documents", "files", "gantt", "issue_tracking", "news", 
"repository", "time_tracking", "wiki")}}

\item{...}{Further parameters}

\item{project_id}{Project id}
}
\value{
id of the created project for \code{redmine_create_project}
}
\description{
These functions implement Redmine API calls to work with projects.
}
\examples{
\dontrun{
  newProjectId <- redmine_create_project(name = "testProject",
      identifier = "test-project", description = "project to testthings",
      enabled_module_names = c("files", "issue_tracking", "repository", 
          "time_tracking", "wiki"))
  
  redmine_update_project(newProjectId, description = "project to test things")
  redmine_delete_project(newProjectId)
}
}
\seealso{
\url{http://www.redmine.org/projects/redmine/wiki/Rest_Projects}

\code{\link{redmine_search_id}} to search id by name
}
\author{
Maxim Nazarov
}
