% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{redmine_delete_watcher}
\alias{redmine_delete_watcher}
\title{Remove watcher from an issue}
\usage{
redmine_delete_watcher(issue_id, user_id)
}
\arguments{
\item{issue_id}{Issue id}

\item{user_id}{User id to remove as a watcher}
}
\description{
Remove watcher from an issue
}
\author{
Maxim Nazarov
}
