% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{redmine_download_attachments}
\alias{redmine_download_attachments}
\title{Download attachments from an issue}
\usage{
redmine_download_attachments(issue_id, path = ".")
}
\arguments{
\item{issue_id}{Issue id}

\item{path}{Path to save files to}
}
\description{
Download attachments from an issue
}
\author{
Maxim Nazarov
}
