% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{redmine_issues}
\alias{redmine_issues}
\title{Show all issues as a data frame}
\usage{
redmine_issues(sort = NULL, issue_id = NULL, project_id = NULL,
  subproject_id = NULL, tracker_id = NULL, status_id = NULL,
  assigned_to_id = NULL, parent_id = NULL, query_id, ...)
}
\arguments{
\item{sort}{Sorting columns}

\item{issue_id}{Issue id(s) to filter by}

\item{project_id}{Project id to filter by}

\item{subproject_id}{Sub-project id to filter by}

\item{tracker_id}{Tracker id}

\item{status_id}{Status id}

\item{assigned_to_id}{user id issues are assigned to}

\item{parent_id}{Parent issue id}

\item{query_id}{Custom query id}

\item{...}{Further arguments}
}
\value{
a data frame
}
\description{
Show all issues as a data frame
}
\examples{
\dontrun{
 redmine_issues()
 redmine_issues(project_id = 1)
 redmine_issues(assigned_to_id = 27)
}
}
\author{
Maxim Nazarov
}
