% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{redmine_list_issues}
\alias{redmine_list_issues}
\title{List issues}
\usage{
redmine_list_issues(offset = NULL, limit = NULL, sort = NULL,
  issue_id = NULL, project_id = NULL, subproject_id = NULL,
  tracker_id = NULL, status_id = NULL, assigned_to_id = NULL,
  parent_id = NULL, ...)
}
\arguments{
\item{offset}{Number of issues to skip}

\item{limit}{Limit number of issues returned}

\item{sort}{Sorting columns}

\item{issue_id}{Issue id(s) to filter by}

\item{project_id}{Project id to filter by}

\item{subproject_id}{Sub-project id to filter by}

\item{tracker_id}{Tracker id}

\item{status_id}{Status id}

\item{assigned_to_id}{user id issues are assigned to}

\item{parent_id}{Parent issue id}

\item{...}{Further arguments}
}
\value{
a \code{redminer} object
}
\description{
Implements Redmine API call to list issues. Note that the output will be 
limited to one page, but \code{offset} or \code{page} can be specified.
}
\examples{
\dontrun{
 redmine_list_issues(project_id = 1, issue_id="123,124")
 redmine_list_issues(assigned_to_id = 27, limit = 5, offset = 200)
 redmine_list_issues(project_id = 2, created_on = ">=2017-06-02T08:12:32Z")
}
}
\seealso{
\code{\link{redmine_issues}} to show all issues as a data frame
}
\author{
Maxim Nazarov
}
