% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{redmine_list_projects}
\alias{redmine_list_projects}
\title{List projects}
\usage{
redmine_list_projects(include = NULL, ...)
}
\arguments{
\item{include}{Which extra info to include, either NULL (no, default), "all" 
or a subset of \code{c("trackers", "issue_categories", "enabled_modules",
"time_entry_activities")}}

\item{...}{further query arguments, such as \code{offset}, \code{limit} or 
\code{page}}
}
\value{
a \code{redminer} object
}
\description{
Implements Redmine API call to list projects. Note that the output will be 
limited to one page, but \code{offset} or \code{page} can be specified.
}
\examples{
\dontrun{
 redmine_list_projects(include = "all")
 redmine_list_projects(offset = 50)
 redmine_list_projects(limit = 20, page = 3)
}
}
\seealso{
\code{\link{redmine_projects}} to show all projects as a data frame
}
\author{
Maxim Nazarov
}
