% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{redmine_move_issue}
\alias{redmine_move_issue}
\title{Move issue to another project}
\usage{
redmine_move_issue(issue_id, new_project_id, notes = NULL, ...)
}
\arguments{
\item{issue_id}{Issue id}

\item{new_project_id}{id of the project to move issue to}

\item{notes}{notes (comments) to add}

\item{...}{extra arguments to passed to \code{\link{redmine_update_issue}}}
}
\description{
A wrapper on top of the generic \code{\link{redmine_update_issue}}
}
\author{
Maxim Nazarov
}
