% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{redmine_projects}
\alias{redmine_projects}
\title{Show all projects as a data frame}
\usage{
redmine_projects(include = NULL)
}
\arguments{
\item{include}{Which extra info to include, either NULL (no, default), "all" 
or a subset of \code{c("trackers", "issue_categories", "enabled_modules")}}
}
\value{
a data frame
}
\description{
Show all projects as a data frame
}
\examples{
\dontrun{
 redmine_projects()
 redmine_projects(include = "all")
}
}
\author{
Maxim Nazarov
}
