% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{redmine_show_issue}
\alias{redmine_show_issue}
\alias{redmine_get_issue}
\title{Show issue information}
\usage{
redmine_show_issue(issue_id, include = NULL)

redmine_get_issue(issue_id, include = NULL)
}
\arguments{
\item{issue_id}{Issue id}

\item{include}{Which extra info to include, either NULL (no, default), "all" 
or a subset of \code{c("children", "attachments", "relations", "changesets", 
      "journals", "watchers")}}
}
\value{
a \code{redminer} object with issue information
}
\description{
Show issue information
}
\examples{
\dontrun{
 issueId <- redmine_search_id("Urgent task", endpoint = "issues")
 redmine_show_issue(issue_id = issueId, include = "all")
}
}
\seealso{
\code{\link{redmine_search_id}} to search for id by subject
}
\author{
Maxim Nazarov
}
