% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{redmine_show_project}
\alias{redmine_show_project}
\alias{redmine_get_project}
\title{Show project information}
\usage{
redmine_show_project(project_id, include = NULL)

redmine_get_project(project_id, include = NULL)
}
\arguments{
\item{project_id}{Project ID}

\item{include}{Which extra info to include, either NULL (no, default), "all" 
or a subset of \code{c("trackers", "issue_categories", "enabled_modules")}}
}
\value{
Project information
}
\description{
Show project information
}
\examples{
\dontrun{
 projectId <- redmine_search_id("testProject", endpoint = "projects")
 redmine_show_project(project_id = projectId, include = "all")
}
}
\seealso{
\code{\link{redmine_search_id}} to search for id by name
}
\author{
Maxim Nazarov
}
