% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{redmine_trackers}
\alias{redmine_trackers}
\alias{redmine_issue_statuses}
\alias{redmine_users}
\alias{redmine_time_entries}
\alias{redmine_issue_priorities}
\alias{redmine_time_entry_activities}
\alias{redmine_document_categories}
\title{Show info on extra endpoints}
\usage{
redmine_trackers()

redmine_issue_statuses()

redmine_users()

redmine_time_entries()

redmine_issue_priorities()

redmine_time_entry_activities()

redmine_document_categories()
}
\value{
a data frame
}
\description{
Convenience functions to list endpoints, in particular to get ids to be 
used in creating/updating or listing issues or projects.
}
\seealso{
\code{\link{redmine_search_id}} to search id by name

\code{\link{redmine_issues}}, \code{\link{redmine_projects}} to 
show info on issues and projects
}
